% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootf.R
\name{bootf}
\alias{bootf}
\title{Garma statistic for tsboot}
\usage{
bootf(data)
}
\arguments{
\item{data}{A univariate or multivariate time series.
It might be vector, matrix or data frame to be passed
 to the \code{garmaFit2} function.?}
}
\description{
This function returns the estimated parameters of a GARMA
model and is used as an input (the \code{statistic}
parameter) for the \code{\link[boot]{tsboot}} function in the \pkg{boot} package.
}
\author{
Matheus de Vasconcellos Barroso
}
