% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dboot.R
\docType{package}
\name{dboot}
\alias{dboot}
\alias{dboot-package}
\title{dboot: Dependent Bootstrap package}
\description{
A package for bootstrapping dependent data. The dboot
package provides two main functions HHJ e LFL
that computes the Moving Block Bootstrap optimal block
length selection algorithms. The former an adapted version
of Hall,
Horowitz and Jing (1995) and the latter Lahiri, Furukawa
and Lee (2007)
}
\section{dboot functions}{

The dboot functions: HHJ and LFL
}

\author{
Matheus de Vasconcellos Barroso
}
