% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_HHJ.R
\docType{data}
\name{example_HHJ}
\alias{example_HHJ}
\title{A realization of a \emph{Poisson-GARMA} model}
\format{A data frame with 30 rows and 4 variables:
\describe{
 \item{indext}{indext, time index of the series}
 \item{mu.t}{mu.t, the observed \eqn{\mu_{t}} value}
 \item{yt}{\eqn{y_{t}}, the realization of the process}
 \item{x}{x, an intercept variable}
 }}
\usage{
example_HHJ
}
\description{
A dataset contaning the realization of a \emph{Poisson-GARMA
(0,1)} model, with \eqn{\theta=0.5, \Beta=2, \mu_{0}=10}
}
\references{
For the \emph{GARMA} model:

Benjamin, Michael A., Rigby, Robert A. and Stasinopoulos, D. Mikis. 2003. Generalized Autoregressive Moving Average Models. Journal of the American Statistical Association. Mar, 2003, Vol. 98, 461, pp. 214-223.
}
\keyword{datasets}
