% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garmaFit2.R
\name{garmaFit2}
\alias{garmaFit2}
\title{A function to fit a GARMA model.}
\usage{
garmaFit2(formula = formula(data), order = c(0, 0), weights = NULL,
  data = sys.parent(), family = NO(), alpha = 0.1, phi.start = NULL,
  theta.start = NULL, tail = max(order), control = list())
}
\description{
This function is for fitting a GARMA model, see Benjamin et al. (2003).
}
\details{
\code{garmaFit2} is a corrected version of the
\code{\link[gamlss.util]{garmaFit}}
function of the \pkg{gamlss.util} package.
}
\note{
For bugs and further requests please refer to
\url{https://github.com/matheusbarroso/dboot}
}
\references{
Benjamin, Michael A., Rigby, Robert A. and
Stasinopoulos, D. Mikis. 2003. Generalized Autoregressive
Moving Average Models. Journal of the American
Statistical Association. Mar, 2003, Vol. 98, 461, pp.
214-223.
}
\author{
Matheus de Vasconcellos Barroso, \email{matheus.vb@gmail.com}
}
