% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.collection.R
\name{M.collection}
\alias{M.collection}
\title{Internal \code{\link{LFL}} function for computing I(i)
in the JAB variance estimator.}
\usage{
M.collection(m, M, R, blocks)
}
\arguments{
\item{m}{The number of bootstrap blocks
to be deleted. The default value is:
\eqn{m = c_{2}*n^(1/3)*l^(2/3)}}

\item{M}{The number of usable observations after
deleting \code{m}, \eqn{M = n-l-m-2}}

\item{R}{A positive integer giving the number of
bootstrap replicates required.}

\item{blocks}{A list with two elements, the first named
starts and the second lenghts. The former is a pxq matrix,
\eqn{p = R} and \eqn{q = ceiling(n/l)} giving the p block
resample initial positions. The latter gives the length of
each block. \code{blocks} is usually equal to
\code{estimate.l$blocks}, the block output of the
 \code{\link{tsboot2}} function.}
}
\value{
A list with the index of reusable blocks of length
\code{M}. This is the I(i) step in the JAB algorithm.
}
\description{
\code{M.collection} computes the subcollection of all
resampled block-sets of size \code{R} where none of the
resampled blocks equals the deleted blocks.
}
\details{
\code{M.collection} is an internal function called by
\code{\link{LFL}}. It has the resampled blocks as inputs
and the reusable blocks as output.
}
\note{
For bugs and further requests please refer to
\url{https://github.com/matheusbarroso/dboot}
}
\references{
Lahiri, S. N. 2002. ON THE JACKKNIFE-AFTER-BOOTSTRAP METHOD
FOR DEPENDENT DATA AND ITS CONSISTENCY PROPERTIES.
Econometric Theory. 18, 2002, pp. 79-98.

Lahiri, Soumendra, Furukawa, Kyoji and Lee, Yd. 2007.
A nonparametric plug-in rule for selecting optimal block
lengths for block bootstrap methods. Statistical
Methodology. July, 2007, Vol. 4, 3, pp. 292-321.

Barroso, Matheus de V. 2018.  BOOTSTRAP METHODS FOR
GENERALIZED AUTOREGRESSIVE MOVING AVERAGE MODELS
}
\author{
Matheus de Vasconcellos Barroso
}
