% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.return.R
\name{ts.return}
\alias{ts.return}
\title{Internal \code{\link{tsboot2}} function.}
\usage{
ts.return(t0, t, R, tseries, seed, stat, n.sim, l, ran.gen, ran.args, call,
  blocks)
}
\arguments{
\item{t0}{\code{t0} is the result of
\code{statistic(tseries,...{})}, as in \pkg{boot}}

\item{t}{The results of applying \code{statistic} to the
replicate time series.}

\item{R}{A positive integer giving the number of
bootstrap replicates required.}

\item{tseries}{The original time series.}

\item{seed}{The used seed in the bootstrap replication.}

\item{stat}{\code{stat =  statistic}, as defined in
\code{\link{tsboot2}}}

\item{n.sim}{The length of the simulated series.}

\item{l}{\code{l} is the fixed block length used in generating the replicate
time series}

\item{ran.gen}{This is a function of three arguments. The first
argument is a time series, it is the result of selecting
\code{n.sim} observations from \code{tseries} by some scheme
and converting the result back into a time series of the
same form as \code{tseries} (with length \code{n.sim}).
The second argument to \code{ran.gen} is always the value
\code{n.sim}, and the third argument is \code{ran.args},
which is used to supply any other objects needed by
\code{ran.gen}.}

\item{ran.args}{This will be supplied to \code{ran.gen}
each time it is called. If \code{ran.gen} needs any
extra arguments then they should be supplied as components
of \code{ran.args}. Multiple arguments may be passed by
making \code{ran.args} a list. If \code{ran.args} is
\code{NULL} then it should not be used within
\code{ran.gen} but note that \code{ran.gen} must still
have its third argument.}

\item{call}{The \code{\link{tsboot2}} call.}

\item{blocks}{Blocks to be returned.}
}
\value{
A list with its arguments.
}
\description{
Function to return the values \code{\link{tsboot2}}
}
\details{
\code{ts.return} is an internal function called by
\code{\link{tsboot2}}. It is a modified version of the
\code{\link[boot]{ts.return}} function of \pkg{boot}.
}
\note{
For bugs and further requests please refer to
\url{https://github.com/matheusbarroso/dboot}
}
\references{

}
\author{
Matheus de Vasconcellos Barroso, \email{matheus.vb@gmail.com}
}
