% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsboot2.R
\name{tsboot2}
\alias{tsboot2}
\title{Internal \code{\link{LFL}} function.}
\usage{
tsboot2(tseries, statistic, R = 100, l = NULL, n.sim = NROW(tseries),
  ran.gen = function(tser, n.sim, args) tser, ran.args = NULL,
  allow.parallel = TRUE, seed = 123, packages = NULL, export = NULL,
  ...)
}
\arguments{
\item{tseries}{A univariate or multivariate time series.}

\item{statistic}{A function which when applied to
\code{tseries} returns a vector containing the
statistic(s) of interest. Each time \code{statistic} is
called it is passed a time series of length \code{n.sim}
which is of the same class as the original \code{tseries}.
Any other arguments which \code{statistic}  takes must
remain constant for each bootstrap replicate and should
be supplied through the \dots argument to
\code{\link{tsboot2}}.}

\item{R}{A positive integer giving the number of
bootstrap replicates required.}

\item{l}{\code{l} is the fixed block length used in generating the replicate
time series}

\item{n.sim}{The length of the simulated series.}

\item{ran.gen}{This is a function of three arguments. The first
argument is a time series, it is the result of selecting
\code{n.sim} observations from \code{tseries} by some scheme
and converting the result back into a time series of the
same form as \code{tseries} (with length \code{n.sim}).
The second argument to \code{ran.gen} is always the value
\code{n.sim}, and the third argument is \code{ran.args},
which is used to supply any other objects needed by
\code{ran.gen}.}

\item{ran.args}{This will be supplied to \code{ran.gen}
each time it is called. If \code{ran.gen} needs any
extra arguments then they should be supplied as components
of \code{ran.args}. Multiple arguments may be passed by
making \code{ran.args} a list. If \code{ran.args} is
\code{NULL} then it should not be used within
\code{ran.gen} but note that \code{ran.gen} must still
have its third argument.}

\item{allow.parallel}{Logical TRUE/FALSE indicating
whether parallel computation via the foreach package
should be used. The default value is TRUE. OBS:paralllel
 backend must be registered prior to calling
 \code{\link{tsboot2}}.}

\item{seed}{The used seed in the bootstrap replication.}

\item{packages}{If \code{allow.parallel = TRUE}.
A character vector with the lisf of
packages required by \code{statisitc}.}

\item{export}{If \code{allow.parallel = TRUE}.
A character vector with the lisf of
objects (functions, etc...) required by \code{statisitc}.}

\item{\dots}{Extra argumetns to \code{statistic} may be
supplied here. Beware of partial matching to the
arguments of \code{\link{tsboot2}} listed above.}

\item{seed}{Numeric, the seed to \code{set.seed()} for
replicable examples.}
}
\value{
An object of class "boot" with similar components
as \code{\link[boot]{tsboot}}
}
\description{
Generate R bootstrap replicates of a statistic applied
to a time series. The replicate time series can be
generated using fixed lengths replicates.
}
\details{
This is a modified version of the
\code{\link[boot]{tsboot}} of the \pkg{boot} package.
Here, only Moving Block Bootstrap is implemented,
the function is modified to return the resampled blocks
and some unecessary arguments were deleted. Additionaly,
the \code{parallel} option is through
the \pkg{doParallel} package.
}
\note{
For bugs and further requests please refer to
\url{https://github.com/matheusbarroso/dboot}
}
\examples{
\dontrun{
library(dboot)
library(gamlss)
library(doParallel)
no_cores <- if(detectCores()==1) 1 else detectCores() -1
registerDoParallel(no_cores)
bootf <- function (db,ord,fam) {
 fit2 <- garmaFit2(yt~x-1,data=db,order=ord,family=fam,tail=0,control=list(iter.max=1000))
 return(fit2$coef)}
ord <- c(1,1) ; fam="GA"
db <- example_LFL
tsboot2(db, statistic=bootf, R = 10, l = 100,ord=ord,fam=fam,export=c("garmaFit2"),package=c("gamlss"))
}

}
\references{
Angelo Canty and Brian Ripley (2017). boot: Bootstrap R
(S-Plus) Functions. R package version 1.3-19.

Davison, A. C. & Hinkley, D. V. (1997) Bootstrap
Methods and Their Applications. Cambridge University
Press, Cambridge. ISBN 0-521-57391-2
}
\author{
Matheus de Vasconcellos Barroso, \email{matheus.vb@gmail.com}
}
