% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsboot2.R
\name{tsboot2}
\alias{tsboot2}
\title{Internal \code{\link{LFL}} function.}
\usage{
tsboot2(tseries, statistic, R, l = NULL, sim = "fixed", endcorr = TRUE,
  n.sim = NROW(tseries), orig.t = TRUE, ran.gen = function(tser, n.sim,
  args) tser, ran.args = NULL, norm = TRUE, ..., parallel = c("no",
  "multicore", "snow"), ncpus = getOption("boot.ncpus", 1L), cl = NULL)
}
\arguments{
\item{tseries}{A univariate or multivariate time series.}

\item{statistic}{A function which when applied to
\code{tseries} returns a vector containing the
statistic(s) of interest. Each time \code{statistic} is
called it is passed a time series of length \code{n.sim}
which is of the same class as the original \code{tseries}.
Any other arguments which \code{statistic}  takes must
remain constant for each bootstrap replicate and should
be supplied through the \dots argument to
\code{\link{tsboot2}}.}

\item{R}{A positive integer giving the number of
bootstrap replicates required.}

\item{l}{\code{l} is the fixed block length used in generating the replicate
time series}

\item{sim}{A character with the 'fixed' tag. (REMOVE the par.)}

\item{endcorr}{\code{FALSE} (REMOVE the par.)}

\item{n.sim}{The length of the simulated series. (REMOVE the par.)}

\item{orig.t}{(REMOVE the par.)}

\item{ran.gen}{This is a function of three arguments. The first
argument is a time series, it is the result of selecting
\code{n.sim} observations from \code{tseries} by some scheme
and converting the result back into a time series of the
same form as \code{tseries} (with length \code{n.sim}).
The second argument to \code{ran.gen} is always the value
\code{n.sim}, and the third argument is \code{ran.args},
which is used to supply any other objects needed by
\code{ran.gen}.}

\item{ran.args}{This will be supplied to \code{ran.gen}
each time it is called. If \code{ran.gen} needs any
extra arguments then they should be supplied as components
of \code{ran.args}. Multiple arguments may be passed by
making \code{ran.args} a list. If \code{ran.args} is
\code{NULL} then it should not be used within
\code{ran.gen} but note that \code{ran.gen} must still
have its third argument.}

\item{norm}{Not used (REMOVE the par.)}

\item{\dots}{Blocks to be returned.}

\item{parallel}{Not used (REMOVE the par.)}

\item{ncpus}{Not used (REMOVE the par.)}

\item{cl}{Not used (REMOVE the par.)}

\item{seed}{The used seed in the bootstrap replication.}

\item{call}{The \code{\link{tsboot2}} call.}
}
\value{
A list with the two elements:
\item{starts }{The initial position of the resampled blocks}
\item{lengths }{The length of the blocks in \code{starts}}
}
\description{
Generate R bootstrap replicates of a statistic applied
to a time series. The replicate time series can be
generated using fixed lengths replicates.
}
\details{
This is a modified version of the
\code{\link[boot]{tsboot}} of the \pkg{boot} package.
Here, only Moving Block Bootstrap is implemented,
the function is modified to return the resampled blocks
and some unecessary arguments were deleted. Additionaly,
the \code{parallel} option is through
the \pkg{doParallel} package.
}
\note{
For bugs and further requests please refer to
\url{https://github.com/matheusbarroso/dboot}
}
\references{
Angelo Canty and Brian Ripley (2017). boot: Bootstrap R
(S-Plus) Functions. R package version 1.3-19.

Davison, A. C. & Hinkley, D. V. (1997) Bootstrap
Methods and Their Applications. Cambridge University
Press, Cambridge. ISBN 0-521-57391-2
}
\author{
Matheus de Vasconcellos Barroso, \email{matheus.vb@gmail.com}
}
